Обновлен фреймворк до 9 версии

+++ app/Http/Controllers/UserController.php
@@ -36,7 +36,7 @@ class UserController extends Controller
 
         $user->update($params);
 
-        if ($path) {
+        if ($path && $oldFile) {
             Storage::disk('public')->delete($oldFile);
         }
